requirements.txt
altgraph==0.17.3
appdirs==1.4.4
APScheduler==3.10.1
asgiref==3.7.2
attrs==23.1.0
autopage==0.5.1
bcrypt==4.0.1
botocore==1.29.160
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cliff==4.3.0
cmd2==2.4.3
cryptography==41.0.1
debtcollector==2.5.0
decorator==5.1.1
Django==4.2.2
dnspython==2.3.0
dogpile.cache==1.2.1
eventlet==0.33.3
fasteners==0.18
fixtures==4.1.0
freezer==14.0.0
greenlet==2.0.2
idna==3.4
image==1.5.33
importlib-metadata==6.7.0
iso8601==2.0.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
keystoneauth1==5.2.1
macholib==1.16.2
msgpack==1.0.5
netaddr==0.8.0
netifaces==0.11.0
openstacksdk==1.3.0
os-brick==6.3.0
os-client-config==2.1.0
os-service-types==1.7.0
os-win==5.9.0
osc-lib==2.8.0
oslo.concurrency==5.1.1
oslo.config==9.1.1
oslo.context==5.1.1
oslo.i18n==6.0.0
oslo.log==5.2.0
oslo.privsep==3.1.0
oslo.serialization==5.1.1
oslo.service==3.1.1
oslo.utils==6.1.0
packaging==23.1
paramiko==3.2.0
Paste==3.5.3
PasteDeploy==3.0.1
pbr==5.11.1
Pillow==9.5.0
prettytable==3.8.0
psutil==5.9.5
pycparser==2.21
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pymongo==4.4.0
PyMySQL==1.0.3
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
pyperclip==1.8.2
pyrsistent==0.19.3
python-cinderclient==9.3.0
python-dateutil==2.8.2
python-freezerclient==5.1.0
python-glanceclient==4.3.0
python-keystoneclient==5.1.0
python-neutronclient==10.0.0
python-novaclient==18.3.0
python-swiftclient==4.3.0
pytz==2023.3
PyYAML==6.0
repoze.lru==0.7
requests==2.31.0
requestsexceptions==1.4.0
rfc3986==2.0.0
Routes==2.5.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stevedore==5.1.0
tenacity==8.2.2
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==1.26.16
warlock==2.0.1
wcwidth==0.2.6
WebOb==1.8.7
wrapt==1.15.0
yappi==1.4.0
zipp==3.15.0
